Move build_request and build_response to utils.
Refs mozilla-services/kinto#71
